+++ src/accessory/DiffuserAccessory.ts
@@ -1,6 +1,6 @@
-import { TuyaDeviceSchemaIntegerProperty, TuyaDeviceStatus } from '../device/TuyaDevice';
-import { limit, remap } from '../util/util';
 import BaseAccessory from './BaseAccessory';
+import { configureActive } from './characteristic/Active';
+import { configureLight } from './characteristic/Light';
 import { configureOn } from './characteristic/On';
 import { configureRotationSpeedLevel } from './characteristic/RotationSpeed';
 
@@ -11,7 +11,7 @@ const SCHEMA_CODE = {
   SPRAY_LEVEL: ['level'],
   LIGHT_ON: ['switch_led'],
   LIGHT_MODE: ['work_mode'],
-  LIGHT_BRIGHTNESS: ['bright_value', 'bright_value_v2'],
+  LIGHT_BRIGHT: ['bright_value', 'bright_value_v2'],
   LIGHT_COLOR: ['colour_data_hsv'],
   SOUND_ON: ['switch_sound'],
 };
@@ -23,9 +23,21 @@ export default class DiffuserAccessory extends BaseAccessory {
   }
 
   configureServices() {
-    configureOn(this, undefined, this.getSchema(...SCHEMA_CODE.ON)); // Main Switch
-    this.configureAirPurifier();
-    this.configureLight();
+    // Main Switch
+    configureOn(this, undefined, this.getSchema(...SCHEMA_CODE.ON));
+
+    this.configureDiffuser();
+
+    configureLight(
+      this,
+      undefined,
+      this.getSchema(...SCHEMA_CODE.LIGHT_ON),
+      this.getSchema(...SCHEMA_CODE.LIGHT_BRIGHT),
+      undefined,
+      this.getSchema(...SCHEMA_CODE.LIGHT_COLOR),
+      this.getSchema(...SCHEMA_CODE.LIGHT_MODE),
+    );
+
     configureOn(this, undefined, this.getSchema(...SCHEMA_CODE.SOUND_ON)); // Sound Switch
   }
 
@@ -34,41 +46,15 @@ export default class DiffuserAccessory extends BaseAccessory {
       || this.accessory.addService(this.Service.AirPurifier);
   }
 
-  lightService() {
-    return this.accessory.getService(this.Service.Lightbulb)
-      || this.accessory.addService(this.Service.Lightbulb, this.device.name + ' Light');
-  }
-
-  configureAirPurifier() {
-    const onSchema = this.getSchema(...SCHEMA_CODE.ON);
+  configureDiffuser() {
     const sprayOnSchema = this.getSchema(...SCHEMA_CODE.SPRAY_ON)!;
 
     // Required Characteristics
-    const { INACTIVE, ACTIVE } = this.Characteristic.Active;
-    this.mainService().getCharacteristic(this.Characteristic.Active)
-      .onGet(() => {
-        if (onSchema && !this.getStatus(onSchema.code)!.value) {
-          return INACTIVE;
-        }
+    configureActive(this, this.mainService(), sprayOnSchema);
 
-        const status = this.getStatus(sprayOnSchema.code)!;
-        return (status.value as boolean) ? ACTIVE : INACTIVE;
-      })
-      .onSet(value => {
-        const commands = [{ code: sprayOnSchema.code, value: (value === ACTIVE) }];
-        if (onSchema && value === ACTIVE) {
-          commands.push({ code: onSchema.code, value: true });
-        }
-        this.sendCommands(commands, true);
-      });
-
-    const { PURIFYING_AIR } = this.Characteristic.CurrentAirPurifierState;
+    const { INACTIVE, PURIFYING_AIR } = this.Characteristic.CurrentAirPurifierState;
     this.mainService().getCharacteristic(this.Characteristic.CurrentAirPurifierState)
       .onGet(() => {
-        if (onSchema && this.getStatus(onSchema.code)!.value !== true) {
-          return INACTIVE;
-        }
-
         const status = this.getStatus(sprayOnSchema.code)!;
         return (status.value as boolean) ? PURIFYING_AIR : INACTIVE;
       });
@@ -82,60 +68,4 @@ export default class DiffuserAccessory extends BaseAccessory {
     configureRotationSpeedLevel(this, this.mainService(), this.getSchema(...SCHEMA_CODE.SPRAY_LEVEL));
   }
 
-  configureLight() {
-    const onSchema = this.getSchema(...SCHEMA_CODE.ON);
-    const lightOnSchema = this.getSchema(...SCHEMA_CODE.LIGHT_ON);
-    if (!lightOnSchema) {
-      return;
-    }
-
-    this.lightService().getCharacteristic(this.Characteristic.On)
-      .onGet(() => {
-        if (onSchema && this.getStatus(onSchema.code)!.value !== true) {
-          return false;
-        }
-
-        const status = this.getStatus(lightOnSchema.code)!;
-        return (status.value as boolean);
-      })
-      .onSet(value => {
-        const commands = [{ code: lightOnSchema.code, value: value as boolean }];
-        if (onSchema && value) {
-          commands.push({ code: onSchema.code, value: true });
-        }
-        this.sendCommands(commands, true);
-      });
-    this.configureLightBrightness();
-  }
-
-  configureLightBrightness() {
-    const schema = this.getSchema(...SCHEMA_CODE.LIGHT_BRIGHTNESS);
-    if (!schema) {
-      return;
-    }
-
-    const lightModeSchema = this.getSchema(...SCHEMA_CODE.LIGHT_MODE);
-
-    const { min, max } = schema.property as TuyaDeviceSchemaIntegerProperty;
-    this.lightService().getCharacteristic(this.Characteristic.Brightness)
-      .onGet(() => {
-        const status = this.getStatus(schema.code)!;
-        const value = Math.round(remap(status.value as number, 0, max, 0, 100));
-        return limit(value, 0, 100);
-      })
-      .onSet(value => {
-        let brightness = Math.round(remap(value as number, 0, 100, 0, max));
-        brightness = limit(brightness, min, max);
-
-        const commands: TuyaDeviceStatus[] = [{
-          code: schema.code,
-          value: brightness,
-        }];
-
-        if (lightModeSchema) {
-          commands.push({ code: lightModeSchema.code, value: 'white' });
-        }
-        this.sendCommands(commands, true);
-      });
-  }
 }