+++ src/accessory/HumidifierAccessory.ts
@@ -4,12 +4,17 @@ import BaseAccessory from './BaseAccessory';
 import { configureActive } from './characteristic/Active';
 import { configureCurrentTemperature } from './characteristic/CurrentTemperature';
 import { configureCurrentRelativeHumidity } from './characteristic/CurrentRelativeHumidity';
+import { configureLight } from './characteristic/Light';
 
 const SCHEMA_CODE = {
   ACTIVE: ['switch'],
   CURRENT_HUMIDITY: ['humidity_current'],
   TARGET_HUMIDITY: ['humidity_set'],
   CURRENT_TEMP: ['temp_current'],
+  LIGHT_ON: ['switch_led'],
+  LIGHT_MODE: ['work_mode'],
+  LIGHT_BRIGHT: ['bright_value', 'bright_value_v2'],
+  LIGHT_COLOR: ['colour_data', 'colour_data_hsv'],
 };
 
 export default class HumidifierAccessory extends BaseAccessory {
@@ -19,13 +24,27 @@ export default class HumidifierAccessory extends BaseAccessory {
   }
 
   configureServices() {
+    // Required Characteristics
     configureActive(this, this.mainService(), this.getSchema(...SCHEMA_CODE.ACTIVE));
-    this.configureTargetState();
     this.configureCurrentState();
+    this.configureTargetState();
     configureCurrentRelativeHumidity(this, this.mainService(), this.getSchema(...SCHEMA_CODE.CURRENT_HUMIDITY));
+
+    // Optional Characteristics
     this.configureRelativeHumidityHumidifierThreshold();
-    configureCurrentTemperature(this, undefined, this.getSchema(...SCHEMA_CODE.CURRENT_TEMP));
     this.configureRotationSpeed();
+
+    // Other
+    configureCurrentTemperature(this, undefined, this.getSchema(...SCHEMA_CODE.CURRENT_TEMP));
+    configureLight(
+      this,
+      undefined,
+      this.getSchema(...SCHEMA_CODE.LIGHT_ON),
+      this.getSchema(...SCHEMA_CODE.LIGHT_BRIGHT),
+      undefined,
+      this.getSchema(...SCHEMA_CODE.LIGHT_COLOR),
+      this.getSchema(...SCHEMA_CODE.LIGHT_MODE),
+    );
   }