+++ CHANGELOG.md
@@ -29,7 +29,10 @@
 - Update support for RGB Power Switch (`dj`).
 - Support showing device online status via `StatusActive`. (#172)
 - Update unit and range of `RotationSpeed` with level, need clean accessory cache to take effect. (#174)
-- Support Door and Window Controller motor reverse mode.
+- Support Door and Window Controller motor reverse mode. (#180)
+- Support Diffuser RGB light. (#184)
+- Support Fan light temperature and color. (#184)
+- Support Humidifier light. (#184)
 
 
 ## [1.6.0] - (2022.12.3)